BasePhonopyWorkChain
BasePhonopyWorkChain was created to separate inputs.calculator_settings.
This aims to make VaspPhonopyWorkChain or QEPhonopyWOrkChain, etc.
PhonopyWorkChain works same as before.
